Error Fixed and Properties Int to Decimal and Null check and Coin Flip Game Page

+++ NumGameWeb/Data/ApplicationUser.cs
@@ -6,7 +6,7 @@ namespace NumGameWeb.Data
         public string? full_name { get; set; }
         public string? email { get; set; }
         public string? phone { get; set; }
-        public int wallet { get; set; }
+        public decimal wallet { get; set; }
         public string? token { get; set; }
     }
