+++ NumGameWeb/WalletResponce.cs
@@ -4,7 +4,7 @@
     {
         public bool status { get; set; }
         public string message { get; set; }
-        public int wallet_balance { get; set; }
+        public decimal wallet_balance { get; set; }
     }
